Main work

+++ BSTtree/Tester.h
@@ -26,7 +26,7 @@ private:
 	size_t _testNum[amountt] = { 10,50,100,200,500,1000,2000,5000,10000,20000,50000,100000,200000,300000,400000,500000};
 	int progress = 0;
 	_times times[10];
-	double time[7][10];
-	void getData(double sum[], double sumo[]);
-	void createFileTemp(double sum[], double sumo[]);
+	long double time[7][10];
+	void getData(long double sum[],long double sumo[]);
+	void createFileTemp(long double sum[],long double sumo[]);
 };
